Refactor to abstract/generic QueryBuilder base
