+++ src/node/ArchivistMongoSdk/index.ts
@@ -1,2 +1,4 @@
+export * from './BoundWitnessQueryBuilder'
 export * from './BoundWitnessSdk'
 export * from './PayloadSdk'
+export * from './QueryBuilder'
